ダイアログ追加

+++ app/src/main/java/com/example/passwordmanager/ShowActivity.kt
@@ -1,15 +1,15 @@
 package com.example.passwordmanager
 
+import android.app.AlertDialog
+import android.content.DialogInterface
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
-import android.widget.Toast
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import io.realm.Realm
-import io.realm.Sort
 import io.realm.kotlin.where
 
 class ShowActivity : AppCompatActivity() {
@@ -18,9 +18,6 @@ class ShowActivity : AppCompatActivity() {
     private lateinit var showAdapter: ShowAdapter
     private lateinit var layoutManager: RecyclerView.LayoutManager
     var id = 0L
-
-    //    val nameList: List<String> = listOf()(unique?.name1,unique?.name2, unique?.name3 , unique?.name4, unique?.name5, unique?.name6)
-//    val passList: Array<String> = arrayOf(unique?.password1 ?: "" , unique?.password2 ?: "", unique?.password3 ?: "", unique?.password4 ?: "", unique?.password5 ?: "", unique?.password6 ?: "")
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -56,14 +53,26 @@ class ShowActivity : AppCompatActivity() {
                 startActivity(intent)
             }
             R.id.menuListOptionDelete -> {
-                realm.executeTransaction {
-                    //一致するidを探して、delete
-                    realm.where<MyModel>()
-                        .equalTo("id", id)
-                        .findFirst()?.deleteFromRealm()
-                }
-                Toast.makeText(applicationContext,"削除しました",Toast.LENGTH_SHORT).show()
-                finish()
+//                val deleteDialog = DeleteDialog(id)
+//                deleteDialog.show(supportFragmentManager, "DeleteDialog")
+                AlertDialog.Builder(this)
+                    .setMessage(R.string.confirm)
+                    .setPositiveButton(R.string.yes,
+                        DialogInterface.OnClickListener { dialog, id ->
+                            realm = Realm.getDefaultInstance()
+                            realm.executeTransaction {
+                                //一致するidを探して、delete
+                                realm.where<MyModel>()
+                                    .equalTo("id", this.id)
+                                    .findFirst()?.deleteFromRealm()
+                            }
+                            finish()
+                        })
+                    .setNegativeButton(R.string.no,
+                        DialogInterface.OnClickListener { dialog, id ->
+
+                        })
+                    .show()
             }
         }
         return super.onOptionsItemSelected(item)
